+++ include/libu8/u8crypto.h
@@ -44,7 +44,7 @@ U8_EXPORT u8_condition u8_CipherInit_Failed;
 U8_EXPORT unsigned char *u8_hmac_sha256
   (const unsigned char *key,int key_len,
    const unsigned char *data,int data_len,
-   unsigned char *result,int *result_len);
+   unsigned char *result,ssize_t *result_len);
 
 
 /** Returns a signed HMAC-SHA384 signature (48 bytes) of a data buffer
@@ -62,7 +62,7 @@ U8_EXPORT unsigned char *u8_hmac_sha256
 U8_EXPORT unsigned char *u8_hmac_sha384
   (const unsigned char *key,int key_len,
    const unsigned char *data,int data_len,
-   unsigned char *result,int *result_len);
+   unsigned char *result,ssize_t *result_len);
 
 /** Returns a signed HMAC-SHA512 signature (64 bytes) of a data buffer
     @param key         a pointer to a key buffer
@@ -79,7 +79,7 @@ U8_EXPORT unsigned char *u8_hmac_sha384
 U8_EXPORT unsigned char *u8_hmac_sha512
   (const unsigned char *key,int key_len,
    const unsigned char *data,int data_len,
-   unsigned char *result,int *result_len);
+   unsigned char *result,ssize_t *result_len);
 
 
 /* Computes the CRC32 code for a buffer.
